+++ src/engin/_assembler.py
@@ -8,7 +8,7 @@ from typing import Any, Generic, TypeVar, cast
 
 from engin._dependency import Dependency, Provide, Supply
 from engin._exceptions import ProviderError
-from engin._type_utils import TypeId, type_id_of
+from engin._type_utils import TypeId
 
 LOG = logging.getLogger("engin")
 
@@ -54,8 +54,7 @@ class Assembler:
     def __init__(self, providers: Iterable[Provide]) -> None:
         self._providers: dict[TypeId, Provide[Any]] = {}
         self._multiproviders: dict[TypeId, list[Provide[list[Any]]]] = defaultdict(list)
-        self._dependencies: dict[TypeId, Any] = {}
-        self._consumed_providers: set[Provide[Any]] = set()
+        self._assembled_outputs: dict[TypeId, Any] = {}
         self._lock = asyncio.Lock()
 
         for provider in providers:
@@ -106,13 +105,14 @@ class Assembler:
         Returns:
             The constructed value.
         """
-        type_id = type_id_of(type_)
-        if type_id in self._dependencies:
-            return cast(T, self._dependencies[type_id])
+        type_id = TypeId.from_type(type_)
+        if type_id in self._assembled_outputs:
+            return cast("T", self._assembled_outputs[type_id])
         if type_id.multi:
-            out = []
             if type_id not in self._multiproviders:
                 raise LookupError(f"no provider found for target type id '{type_id}'")
+
+            out = []
             for provider in self._multiproviders[type_id]:
                 assembled_dependency = await self.assemble(provider)
                 try:
@@ -123,11 +123,12 @@ class Assembler:
                         error_type=type(err),
                         error_message=str(err),
                     ) from err
-            self._dependencies[type_id] = out
+            self._assembled_outputs[type_id] = out
             return out  # type: ignore[return-value]
         else:
             if type_id not in self._providers:
                 raise LookupError(f"no provider found for target type id '{type_id}'")
+
             assembled_dependency = await self.assemble(self._providers[type_id])
             try:
                 value = await assembled_dependency()
@@ -137,7 +138,7 @@ class Assembler:
                     error_type=type(err),
                     error_message=str(err),
                 ) from err
-            self._dependencies[type_id] = value
+            self._assembled_outputs[type_id] = value
             return value  # type: ignore[return-value]
 
     def has(self, type_: type[T]) -> bool:
@@ -150,7 +151,7 @@ class Assembler:
         Returns:
             True if the Assembler has a provider for type else False.
         """
-        type_id = type_id_of(type_)
+        type_id = TypeId.from_type(type_)
         if type_id.multi:
             return type_id in self._multiproviders
         else:
@@ -160,24 +161,26 @@ class Assembler:
         """
         Add a provider to the Assembler post-initialisation.
 
+        If this replaces an existing provider, this will clear any previously assembled
+        output for the existing Provider.
+
         Args:
             provider: the Provide instance to add.
 
         Returns:
              None
-
-        Raises:
-            ValueError: if a provider for this type already exists.
         """
         type_id = provider.return_type_id
         if provider.is_multiprovider:
             if type_id in self._multiproviders:
+                if type_id in self._assembled_outputs:
+                    del self._assembled_outputs[type_id]
                 self._multiproviders[type_id].append(provider)
             else:
                 self._multiproviders[type_id] = [provider]
         else:
-            if type_id in self._providers:
-                raise ValueError(f"A provider for '{type_id}' already exists")
+            if type_id in self._assembled_outputs:
+                del self._assembled_outputs[type_id]
             self._providers[type_id] = provider
 
     def _resolve_providers(self, type_id: TypeId) -> Collection[Provide]:
@@ -192,7 +195,9 @@ class Assembler:
                 # store default to prevent the warning appearing multiple times
                 self._multiproviders[type_id] = providers
             else:
-                raise LookupError(f"No Provider registered for dependency '{type_id}'")
+                available = sorted(str(k) for k in self._providers)
+                msg = f"Missing Provider for type '{type_id}', available: {available}"
+                raise LookupError(msg)
 
         required_providers: list[Provide[Any]] = []
         for provider in providers:
@@ -206,9 +211,11 @@ class Assembler:
 
     async def _satisfy(self, target: TypeId) -> None:
         for provider in self._resolve_providers(target):
-            if provider in self._consumed_providers:
+            if (
+                not provider.is_multiprovider
+                and provider.return_type_id in self._assembled_outputs
+            ):
                 continue
-            self._consumed_providers.add(provider)
             type_id = provider.return_type_id
             bound_args = await self._bind_arguments(provider.signature)
             try:
@@ -218,12 +225,12 @@ class Assembler:
                     provider=provider, error_type=type(err), error_message=str(err)
                 ) from err
             if provider.is_multiprovider:
-                if type_id in self._dependencies:
-                    self._dependencies[type_id].extend(value)
+                if type_id in self._assembled_outputs:
+                    self._assembled_outputs[type_id].extend(value)
                 else:
-                    self._dependencies[type_id] = value
+                    self._assembled_outputs[type_id] = value
             else:
-                self._dependencies[type_id] = value
+                self._assembled_outputs[type_id] = value
 
     async def _bind_arguments(self, signature: Signature) -> BoundArguments:
         args = []
@@ -232,11 +239,11 @@ class Assembler:
             if param_name == "self":
                 args.append(object())
                 continue
-            param_key = type_id_of(param.annotation)
-            has_dependency = param_key in self._dependencies
+            param_key = TypeId.from_type(param.annotation)
+            has_dependency = param_key in self._assembled_outputs
             if not has_dependency:
                 await self._satisfy(param_key)
-            val = self._dependencies[param_key]
+            val = self._assembled_outputs[param_key]
             if param.kind == param.POSITIONAL_ONLY:
                 args.append(val)
             else: