+++ src/engin/_cli/_graph.py
@@ -1,63 +1,72 @@
+import contextlib
 import importlib
 import logging
 import socketserver
 import sys
 import threading
-from argparse import ArgumentParser
 from http.server import BaseHTTPRequestHandler
 from time import sleep
-from typing import Any
+from typing import Annotated, Any
+
+import typer
+from rich import print
 
 from engin import Engin, Entrypoint, Invoke
+from engin._cli._utils import print_error
 from engin._dependency import Dependency, Provide, Supply
 from engin.ext.asgi import ASGIEngin
 from engin.ext.fastapi import APIRouteDependency
 
+cli = typer.Typer()
+
+
 # mute logging from importing of files + engin's debug logging.
 logging.disable()
 
-args = ArgumentParser(
-    prog="engin-graph",
-    description="Creates a visualisation of your application's dependencies",
-)
-args.add_argument(
-    "app",
-    help=(
-        "the import path of your Engin instance, in the form "
-        "'package:application', e.g. 'app.main:engin'"
-    ),
-)
 
 _APP_ORIGIN = ""
 
+_CLI_HELP = {
+    "app": (
+        "The import path of your Engin instance, in the form 'package:application'"
+        ", e.g. 'app.main:engin'"
+    )
+}
 
-def serve_graph() -> None:
+
+@cli.command(name="graph")
+def serve_graph(
+    app: Annotated[
+        str,
+        typer.Argument(help=_CLI_HELP["app"]),
+    ],
+) -> None:
+    """
+    Creates a visualisation of your application's dependencies.
+    """
     # add cwd to path to enable local package imports
     sys.path.insert(0, "")
-
-    parsed = args.parse_args()
-
-    app = parsed.app
 
     try:
         module_name, engin_name = app.split(":", maxsplit=1)
     except ValueError:
-        raise ValueError(
-            "Expected an argument of the form 'module:attribute', e.g. 'myapp:engin'"
-        ) from None
+        print_error("Expected an argument of the form 'module:attribute', e.g. 'myapp:engin'")
 
     global _APP_ORIGIN
     _APP_ORIGIN = module_name.split(".", maxsplit=1)[0]
 
-    module = importlib.import_module(module_name)
+    try:
+        module = importlib.import_module(module_name)
+    except ModuleNotFoundError:
+        print_error(f"unable to find module '{module_name}'")
 
     try:
         instance = getattr(module, engin_name)
-    except LookupError:
-        raise LookupError(f"Module '{module_name}' has no attribute '{engin_name}'") from None
+    except AttributeError:
+        print_error(f"module '{module_name}' has no attribute '{engin_name}'")
 
     if not isinstance(instance, Engin):
-        raise TypeError(f"'{app}' is not an Engin instance")
+        print_error(f"'{app}' is not an Engin instance")
 
     nodes = instance.graph()
 
@@ -96,10 +105,14 @@ def serve_graph() -> None:
     server_thread.daemon = True  # Daemonize the thread so it exits when the main script exits
     server_thread.start()
 
-    try:
-        sleep(10000)
-    except KeyboardInterrupt:
-        print("Exiting the server...")
+    with contextlib.suppress(KeyboardInterrupt):
+        wait_for_interrupt()
+
+    print("Exiting the server...")
+
+
+def wait_for_interrupt() -> None:
+    sleep(10000)
 
 
 _BLOCK_IDX: dict[str, int] = {}