+++ src/engin/_engin.py
@@ -3,26 +3,20 @@ import logging
 import os
 import signal
 from asyncio import Event, Task
-from collections.abc import Iterable
 from contextlib import AsyncExitStack
 from itertools import chain
 from types import FrameType
-from typing import ClassVar, TypeAlias
+from typing import ClassVar
 
-from engin import Entrypoint
 from engin._assembler import AssembledDependency, Assembler
-from engin._block import Block
-from engin._dependency import Dependency, Invoke, Provide, Supply
+from engin._dependency import Invoke, Provide, Supply
 from engin._graph import DependencyGrapher, Node
 from engin._lifecycle import Lifecycle
+from engin._option import Option
 from engin._type_utils import TypeId
 
-LOG = logging.getLogger("engin")
-
-Option: TypeAlias = Invoke | Provide | Supply | Block
-_Opt: TypeAlias = Invoke | Provide | Supply
-
 _OS_IS_WINDOWS = os.name == "nt"
+LOG = logging.getLogger("engin")
 
 
 class Engin:
@@ -88,12 +82,16 @@ class Engin:
         self._shutdown_task: Task | None = None
         self._run_task: Task | None = None
 
-        # TODO: refactor _destruct_options and related attributes into a dedicated class?
-        self._providers: dict[TypeId, Provide] = {TypeId.from_type(Engin): Provide(self._self)}
+        self._providers: dict[TypeId, Provide] = {
+            TypeId.from_type(Engin): Supply(self, type_hint=Engin)
+        }
         self._multiproviders: dict[TypeId, list[Provide]] = {}
         self._invocations: list[Invoke] = []
+
         # populates the above
-        self._destruct_options(chain(self._LIB_OPTIONS, options))
+        for option in chain(self._LIB_OPTIONS, options):
+            option.apply(self)
+
         multi_providers = [p for multi in self._multiproviders.values() for p in multi]
         self._assembler = Assembler(chain(self._providers.values(), multi_providers))
 
@@ -174,46 +172,6 @@ class Engin:
     async def _shutdown_when_stopped(self) -> None:
         await self._stop_requested_event.wait()
         await self._shutdown()
-
-    def _destruct_options(self, options: Iterable[Option]) -> None:
-        for opt in options:
-            if isinstance(opt, Block):
-                self._destruct_options(opt)
-            if isinstance(opt, Provide | Supply):
-                if not opt.is_multiprovider:
-                    existing = self._providers.get(opt.return_type_id)
-                    self._log_option(opt, overwrites=existing)
-                    self._providers[opt.return_type_id] = opt
-                else:
-                    self._log_option(opt)
-                    if opt.return_type_id in self._multiproviders:
-                        self._multiproviders[opt.return_type_id].append(opt)
-                    else:
-                        self._multiproviders[opt.return_type_id] = [opt]
-            elif isinstance(opt, Invoke):
-                self._log_option(opt)
-                self._invocations.append(opt)
-
-    @staticmethod
-    def _log_option(opt: Dependency, overwrites: Dependency | None = None) -> None:
-        if overwrites is not None:
-            extra = f"\tOVERWRITES {overwrites.name}"
-            if overwrites.block_name:
-                extra += f" [{overwrites.block_name}]"
-        else:
-            extra = ""
-        if isinstance(opt, Supply):
-            LOG.debug(f"SUPPLY      {opt.return_type_id!s:<35}{extra}")
-        elif isinstance(opt, Provide):
-            LOG.debug(f"PROVIDE     {opt.return_type_id!s:<35} <- {opt.name}() {extra}")
-        elif isinstance(opt, Entrypoint):
-            type_id = opt.parameter_types[0]
-            LOG.debug(f"ENTRYPOINT  {type_id!s:<35}")
-        elif isinstance(opt, Invoke):
-            LOG.debug(f"INVOKE      {opt.name:<35}")
-
-    def _self(self) -> "Engin":
-        return self
 
 
 async def _wait_for_stop_signal(stop_requested_event: Event) -> None: