+++ src/engin/_lifecycle.py
@@ -1,12 +1,15 @@
 import asyncio
 import logging
+from collections.abc import Awaitable, Callable
 from contextlib import AbstractAsyncContextManager, AbstractContextManager
+from inspect import iscoroutinefunction
 from types import TracebackType
 from typing import TypeAlias, TypeGuard, cast
 
 LOG = logging.getLogger("engin")
 
 _AnyContextManager: TypeAlias = AbstractAsyncContextManager | AbstractContextManager
+_ParameterlessCallable: TypeAlias = Callable[[], None | Awaitable[None]]
 
 
 class Lifecycle:
@@ -44,6 +47,18 @@ class Lifecycle:
 
             lifecycle.append(task)
         ```
+
+        Defining a custom lifecycle using a LifecycleHook.
+
+        ```python
+        def my_provider(lifecycle: Lifecycle) -> str:
+            connection_pool = ConnectionPool()
+
+            lifecycle.hook(
+                on_start=connection_pool.connect,
+                on_stop=connection_pool.close,
+            )
+        ```
     """
 
     def __init__(self) -> None:
@@ -59,6 +74,25 @@ class Lifecycle:
         suppressed_cm = _AExitSuppressingAsyncContextManager(cm)
         self._context_managers.append(suppressed_cm)
 
+    def hook(
+        self,
+        *,
+        on_start: _ParameterlessCallable | None = None,
+        on_stop: _ParameterlessCallable | None = None,
+    ) -> None:
+        """
+        Append a hook to the Lifecycle.
+
+        At least one of `on_start` or `on_stop` must be provided.
+
+        Args:
+            on_start: a callable to be executed on Lifecycle startup.
+            on_stop: a callable to be executed on Lifecycle shutdown.
+        """
+        if on_start is None and on_stop is None:
+            raise ValueError("At least one of on_start or on_stop must be provided")
+        self.append(LifecycleHook(on_start=on_start, on_stop=on_stop))
+
     def list(self) -> list[AbstractAsyncContextManager]:
         """
         List all the defined tasks.
@@ -69,6 +103,38 @@ class Lifecycle:
         return self._context_managers[:]
 
 
+class LifecycleHook(AbstractAsyncContextManager):
+    def __init__(
+        self,
+        on_start: _ParameterlessCallable | None = None,
+        on_stop: _ParameterlessCallable | None = None,
+    ) -> None:
+        self._on_start = on_start
+        self._on_stop = on_stop
+
+    async def __aenter__(self) -> None:
+        if self._on_start is not None:
+            func = self._on_start
+            if iscoroutinefunction(func):
+                await func()
+            else:
+                await asyncio.to_thread(func)
+
+    async def __aexit__(
+        self,
+        exc_type: type[BaseException] | None,
+        exc_value: BaseException | None,
+        traceback: TracebackType | None,
+        /,
+    ) -> None:
+        if self._on_stop is not None:
+            func = self._on_stop
+            if iscoroutinefunction(func):
+                await func()
+            else:
+                await asyncio.to_thread(func)
+
+
 class _AExitSuppressingAsyncContextManager(AbstractAsyncContextManager):
     def __init__(self, cm: _AnyContextManager) -> None:
         self._cm = cm
@@ -77,7 +143,7 @@ class _AExitSuppressingAsyncContextManager(AbstractAsyncContextManager):
         if self._is_async_cm(self._cm):
             await self._cm.__aenter__()
         else:
-            await asyncio.to_thread(cast(AbstractContextManager, self._cm).__enter__)
+            await asyncio.to_thread(cast("AbstractContextManager", self._cm).__enter__)
 
     async def __aexit__(
         self,
@@ -91,7 +157,7 @@ class _AExitSuppressingAsyncContextManager(AbstractAsyncContextManager):
                 await self._cm.__aexit__(exc_type, exc_value, traceback)
             else:
                 await asyncio.to_thread(
-                    cast(AbstractContextManager, self._cm).__exit__,
+                    cast("AbstractContextManager", self._cm).__exit__,
                     exc_type,
                     exc_value,
                     traceback,