+++ src/engin/_type_utils.py
@@ -1,12 +1,16 @@
+import re
 import typing
 from dataclasses import dataclass
+from linecache import getline
 from types import UnionType
 from typing import Any
+
+from engin._introspect import get_first_external_frame
 
 _implict_modules = ["builtins", "typing", "collections.abc", "types"]
 
 
-@dataclass(frozen=True, eq=True, slots=True)
+@dataclass(frozen=True, slots=True)
 class TypeId:
     """
     Represents information about a Type in the Dependency Injection framework.
@@ -14,6 +18,7 @@ class TypeId:
 
     type: type
     multi: bool
+    alias: str | None
 
     @classmethod
     def from_type(cls, type_: Any) -> "TypeId":
@@ -26,11 +31,12 @@ class TypeId:
         Returns:
             The corresponding TypeId for that type.
         """
-        if is_multi_type(type_):
+        alias = _extract_alias(type_)
+        if _is_multi_type(type_):
             inner_obj = typing.get_args(type_)[0]
-            return TypeId(type=inner_obj, multi=True)
+            return TypeId(type=inner_obj, multi=True, alias=alias)
         else:
-            return TypeId(type=type_, multi=False)
+            return TypeId(type=type_, multi=False, alias=alias)
 
     def __str__(self) -> str:
         module = self.type.__module__
@@ -38,7 +44,14 @@ class TypeId:
         out += _args_to_str(self.type)
         if self.multi:
             out += "[]"
+        if self.alias:
+            return f"Alias[{self.alias}, {out}]"
         return out
+
+    def __eq__(self, other: Any) -> bool:
+        if not isinstance(other, TypeId):
+            return False
+        return self.type == other.type and self.multi == other.multi
 
 
 def _args_to_str(type_: Any) -> str:
@@ -65,15 +78,22 @@ def _args_to_str(type_: Any) -> str:
     return arg_str
 
 
-def type_id_of(type_: Any) -> TypeId:
-    """
-    Generates a string TypeId for any type.
-    """
-    return TypeId.from_type(type_)
-
-
-def is_multi_type(type_: Any) -> bool:
+def _is_multi_type(type_: Any) -> bool:
     """
     Discriminates a type to determine whether it is the return type of a multiprovider.
     """
     return typing.get_origin(type_) is list
+
+
+_ALIAS_PATTERN = re.compile(r"from_type\((\w+)\)")
+
+
+def _extract_alias(expected_type: type) -> str | None:
+    source_frame = get_first_external_frame()
+    source_code = getline(source_frame.filename, source_frame.lineno)
+
+    if alias_match := re.search(_ALIAS_PATTERN, source_code):
+        alias = alias_match.group(1)
+        if source_frame.frame.f_locals.get(alias) == expected_type:
+            return alias
+    return None