+++ tests/test_utils.py
@@ -1,69 +1,69 @@
 from collections.abc import Callable
 from typing import Annotated, TypeAlias
 
-from engin._type_utils import TypeId, type_id_of
+from engin._type_utils import TypeId
 
 
 def test_type_id_of_int():
-    type_id = type_id_of(int)
-    assert type_id == TypeId(type=int, multi=False)
+    type_id = TypeId.from_type(int)
+    assert type_id == TypeId(type=int, multi=False, alias=None)
     assert str(type_id) == "int"
 
 
 def test_type_id_of_list_of_int():
-    type_id = type_id_of(list[int])
-    assert type_id == TypeId(type=int, multi=True)
+    type_id = TypeId.from_type(list[int])
+    assert type_id == TypeId(type=int, multi=True, alias=None)
     assert str(type_id) == "int[]"
 
 
 def test_type_id_of_alias():
     MyAlias: TypeAlias = int
 
+    type_id = TypeId.from_type(MyAlias)
+
     assert (
-        type_id_of(MyAlias)
-        == TypeId(type=MyAlias, multi=False)
-        == TypeId(type=int, multi=False)
+        type_id
+        == TypeId(type=MyAlias, multi=False, alias="MyAlias")
+        == TypeId(type=int, multi=False, alias=None)
     )
+    assert str(type_id) == "Alias[MyAlias, int]"
 
 
 def test_type_id_of_type_of_class():
     class Foo: ...
 
-    type_id = type_id_of(type[Foo])
-    assert type_id == TypeId(type=type[Foo], multi=False)
+    type_id = TypeId.from_type(type[Foo])
+    assert type_id == TypeId(type=type[Foo], multi=False, alias=None)
     assert str(type_id) == "type[Foo]"
 
 
 def test_type_id_of_callable():
-    func = Callable[[int], str]
-    type_id = type_id_of(func)
-    assert type_id == TypeId(type=Callable[[int], str], multi=False)
+    type_id = TypeId.from_type(Callable[[int], str])
+    assert type_id == TypeId(type=Callable[[int], str], multi=False, alias=None)
     assert str(type_id) == "Callable[[int], str]"
 
 
 def test_type_id_of_multi_callable():
-    many_func = list[Callable[[int], str]]
-    type_id = type_id_of(many_func)
-    assert type_id == TypeId(type=Callable[[int], str], multi=True)
+    type_id = TypeId.from_type(list[Callable[[int], str]])
+    assert type_id == TypeId(type=Callable[[int], str], multi=True, alias=None)
     assert str(type_id) == "Callable[[int], str][]"
 
 
 def test_type_id_of_annotation():
-    annotated = Annotated[int, "Activity"]
-    type_id = type_id_of(annotated)
-    assert type_id == TypeId(type=Annotated[int, "Activity"], multi=False)
+    type_id = TypeId.from_type(Annotated[int, "Activity"])
+    assert type_id == TypeId(type=Annotated[int, "Activity"], multi=False, alias=None)
     assert str(type_id) == "Annotated[int, Activity]"
 
 
 def test_type_id_of_complex_annotation():
-    annotated = Annotated[Callable[[int], str], "Activity"]
-    type_id = type_id_of(annotated)
-    assert type_id == TypeId(type=Annotated[Callable[[int], str], "Activity"], multi=False)
+    type_id = TypeId.from_type(Annotated[Callable[[int], str], "Activity"])
+    assert type_id == TypeId(
+        type=Annotated[Callable[[int], str], "Activity"], multi=False, alias=None
+    )
     assert str(type_id) == "Annotated[Callable[[int], str], Activity]"
 
 
 def test_type_id_of_union():
-    union = int | str
-    type_id = type_id_of(union)
-    assert type_id == TypeId(type=int | str, multi=False)
+    type_id = TypeId.from_type(int | str)
+    assert type_id == TypeId(type=int | str, multi=False, alias=None)
     assert str(type_id) == "Union[int, str]"